+++ esp-hal/src/dma/pdma.rs
@@ -366,6 +366,15 @@ macro_rules! ImplSpiChannel {
                 }
             }
 
+            impl DmaChannelConvert<AnySpiDmaChannel> for [<Spi $num DmaChannel>] {
+                fn degrade_rx(rx: SpiDmaRxChannelImpl<Self>) -> SpiDmaRxChannelImpl<AnySpiDmaChannelInner> {
+                    SpiDmaRxChannelImpl(rx.0.into())
+                }
+                fn degrade_tx(tx: SpiDmaTxChannelImpl<Self>) -> SpiDmaTxChannelImpl<AnySpiDmaChannelInner> {
+                    SpiDmaTxChannelImpl(tx.0.into())
+                }
+            }
+
             impl $crate::private::Sealed for [<Spi $num DmaChannel>] {}
 
             #[doc = concat!("Creates a channel for SPI", $num)]
@@ -895,5 +904,41 @@ where
             "This DMA channel is not compatible with {:?}",
             peripheral.peripheral()
         );
+    }
+}
+
+/// A marker for SPI-compatible type-erased DMA channels.
+pub struct AnySpiDmaChannel;
+
+impl crate::private::Sealed for AnySpiDmaChannel {}
+
+impl DmaChannel for AnySpiDmaChannel {
+    type Rx = SpiDmaRxChannelImpl<AnySpiDmaChannelInner>;
+    type Tx = SpiDmaTxChannelImpl<AnySpiDmaChannelInner>;
+}
+
+crate::any_enum! {
+    #[doc(hidden)]
+    pub enum AnySpiDmaChannelInner {
+        Spi2(Spi2DmaChannel),
+        Spi3(Spi3DmaChannel),
+    }
+}
+
+impl crate::private::Sealed for AnySpiDmaChannelInner {}
+
+impl PdmaChannel for AnySpiDmaChannelInner {
+    type RegisterBlock = SpiRegisterBlock;
+
+    delegate::delegate! {
+        to match self {
+            AnySpiDmaChannelInner::Spi2(channel) => channel,
+            AnySpiDmaChannelInner::Spi3(channel) => channel,
+        } {
+            fn register_block(&self) -> &SpiRegisterBlock;
+            fn tx_waker(&self) -> &'static AtomicWaker;
+            fn rx_waker(&self) -> &'static AtomicWaker;
+            fn is_compatible_with(&self, peripheral: &impl PeripheralMarker) -> bool;
+        }
     }
 }